Move vsock logic to dedicated package
As part of this move the newCIDsMap needed to
be exported.
