+++ pkg/virt-controller/watch/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
         "replicaset.go",
         "vm.go",
         "vmi.go",
-        "vsock.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-controller/watch",
     visibility = ["//visibility:public"],
@@ -57,6 +56,7 @@ go_library(
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//pkg/virt-controller/watch/volume-migration:go_default_library",
+        "//pkg/virt-controller/watch/vsock:go_default_library",
         "//pkg/virt-controller/watch/workload-updater:go_default_library",
         "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
@@ -118,7 +118,6 @@ go_test(
         "updatereactor_test.go",
         "vm_test.go",
         "vmi_test.go",
-        "vsock_test.go",
         "watch_suite_test.go",
     ],
     embed = [":go_default_library"],
@@ -163,7 +162,6 @@ go_test(
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//staging/src/kubevirt.io/client-go/testing:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
-        "//tests/libvmifact:go_default_library",
         "//vendor/github.com/emicklei/go-restful/v3:go_default_library",
         "//vendor/github.com/evanphx/json-patch:go_default_library",
         "//vendor/github.com/golang/mock/gomock:go_default_library",
